# openapi.yaml
openapi: 3.0.3
info:
  title: MySelfServe API
  version: 0.1.0
  description: |
    BC Government Income Assistance Self-Service Portal API.

    MySS 2.0 provides self-service capabilities for BC income assistance clients
    to manage service requests, monthly reports, payments, employment plans,
    messages, and account settings. Ministry workers can access client data
    through an admin support view.

    ## Authentication

    Most endpoints require a JWT Bearer token issued via BCeID/IDIR through
    Keycloak OIDC. The token is signed with HS256 and contains claims for
    `sub` (user_id), `role`, `bceid_guid`, and `idir_username`.

    ## Roles

    - **CLIENT** — BCeID-authenticated income assistance clients
    - **WORKER** — IDIR-authenticated ministry staff (includes ADMIN)
    - **ADMIN** — System administrators (inherits WORKER permissions)

    ## Error Handling

    Errors from the upstream Siebel/ICM system are mapped to standard HTTP
    status codes with structured JSON error bodies. A circuit breaker protects
    the application during ICM outages (5 failures → 30s open window → 503).
  contact:
    name: MySS Development Team
  license:
    name: BC Government
    url: https://www2.gov.bc.ca/
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://myss-api-dev.apps.silver.devops.gov.bc.ca
    description: Development (OpenShift)
  - url: https://myss-api-test.apps.silver.devops.gov.bc.ca
...
security:
  - BearerAuth: []
paths:
...
  /registration/start:
    post:
      operationId: startRegistration
      tags: [Registration]
      summary: Start a new registration session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRegistrationRequest"
      responses:
        "201":
          description: Registration session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartRegistrationResponse"
  /registration/verify/{invite_token}:
    get:
      operationId: verifyInviteToken
      tags: [Registration]
      summary: Verify an email invitation token
      description: Validates the invite token and redirects to the PIN creation step.
      security: []
      parameters:
        - name: invite_token
          in: path
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to PIN creation step
          headers:
            Location:
              schema:
                type: string
                example: /registration/{session_token}/pin
        "410":
          description: Token expired, used, or not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registration/{token}/step/{step_number}:
    get:
      operationId: getRegistrationStep
      tags: [Registration]
      summary: Get current registration step state
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: step_number
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Step state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepStateResponse"
  /registration/{token}/personal-info:
    post:
      operationId: submitPersonalInfo
      tags: [Registration]
      summary: Submit personal information (step 2)
      description: |
        Submits the applicant's personal information including name, SIN, DOB,
        contact details, and optional spouse information. SIN is validated
        via Luhn algorithm; PHN via MOD 11; age must be >= 16.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalInfoRequest"
      responses:
        "201":
          description: Personal info accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalInfoResponse"
        "409":
          description: Previous steps not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /registration/{token}/pin:
    post:
      operationId: createRegistrationPin
      tags: [Registration]
      summary: Create PIN (step 4)
      description: Sets the client's 4-digit PIN after email verification.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinRequest"
      responses:
        "201":
          description: PIN created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinResponse"
        "409":
          description: Email verification not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registration/{token}/link-bceid:
    post:
      operationId: linkBceid
      tags: [Registration]
      summary: Link BCeID account (step 5)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: BCeID link pending
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: pending
                  message:
                    type: string
                    example: BCeID link not yet implemented
        "409":
          description: Previous steps not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──────────────────────────────────────────────
  # Service Requests
  # ──────────────────────────────────────────────
...
  /admin/ao/registration/{token}/step/{step}:
    get:
      operationId: getAORegistrationStep
      tags: ["Admin: AO Registration"]
      summary: Get AO registration step data
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: step
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Step data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AORegistrationStep"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: submitAORegistrationStep
      tags: ["Admin: AO Registration"]
      summary: Submit AO registration step data
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: step
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Step submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AORegistrationStep"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token signed with HS256. Claims: `sub` (user_id), `role` (CLIENT/WORKER/ADMIN),
        `bceid_guid` (clients), `idir_username` (workers/admins), `exp` (expiry).
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: HMAC secret for internal AV scanner webhook
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
      required:
        - detail
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Field location path
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type
            required:
              - loc
              - msg
              - type

    # ── Health ──────────────────────────────────
...
    AccountCreationType:
      type: string
      enum: [SELF, WITH_HELPER, POA, PARENT]
    Gender:
      type: string
      enum: [M, F, U]
    PhoneType:
      type: string
      enum: [HOME, CELL, WORK, MESSAGE]
...
    StartRegistrationRequest:
      type: object
      required:
        - account_creation_type
      properties:
        account_creation_type:
          $ref: "#/components/schemas/AccountCreationType"
    StartRegistrationResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Session token for subsequent registration steps
    StepStateResponse:
      type: object
      required:
        - token
        - step_reached
        - account_creation_type
      properties:
        token:
          type: string
        step_reached:
          type: integer
        account_creation_type:
          $ref: "#/components/schemas/AccountCreationType"
    OpenCaseIdentificationRequest:
      type: object
      properties:
        last_assistance_payment_amount:
          type: string
          nullable: true
        has_drivers_licence:
          type: boolean
          default: false
        drivers_licence_expiry:
          type: string
          format: date
          nullable: true
        has_previous_address:
          type: boolean
          default: false
        previous_address:
          type: string
          nullable: true
        paying_service_providers:
          type: boolean
          default: false
        service_providers_description:
          type: string
          nullable: true
        payment_method:
          type: string
          nullable: true
    SpouseInfoRequest:
      type: object
      required:
        - first_name
        - last_name
        - sin
        - date_of_birth
        - gender
      properties:
        first_name:
          type: string
        middle_name:
          type: string
          nullable: true
        last_name:
          type: string
        sin:
          type: string
          description: Social Insurance Number (Luhn validated)
        phn:
          type: string
          nullable: true
          description: Personal Health Number (MOD 11 validated)
        date_of_birth:
          type: string
          format: date
          description: Must be at least 16 years old
        gender:
          $ref: "#/components/schemas/Gender"
    PersonalInfoRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - email_confirm
        - sin
        - date_of_birth
        - gender
        - phone_number
        - phone_type
      properties:
        first_name:
          type: string
        middle_name:
          type: string
          nullable: true
        last_name:
          type: string
        email:
          type: string
          format: email
        email_confirm:
          type: string
          format: email
          description: Must match email field
        sin:
          type: string
          description: Social Insurance Number (Luhn validated)
        phn:
          type: string
          nullable: true
          description: Personal Health Number (MOD 11 validated)
        date_of_birth:
          type: string
          format: date
          description: Must be at least 16 years old
        gender:
          $ref: "#/components/schemas/Gender"
        phone_number:
          type: string
        phone_type:
          $ref: "#/components/schemas/PhoneType"
        has_open_case:
          type: boolean
          default: false
        open_case_identification:
          $ref: "#/components/schemas/OpenCaseIdentificationRequest"
        spouse:
          $ref: "#/components/schemas/SpouseInfoRequest"
    PersonalInfoResponse:
      type: object
      required:
        - next_step
      properties:
        next_step:
          type: integer
          example: 3
    PinRequest:
      type: object
      required:
        - pin
        - pin_confirm
      properties:
        pin:
          type: string
          pattern: "^\\d{4}$"
          description: 4-digit PIN (0000-9999)
        pin_confirm:
          type: string
          pattern: "^\\d{4}$"
          description: Must match pin field
    PinResponse:
      type: object
      required:
        - next_step
      properties:
        next_step:
          type: integer
          example: 5

    # ── Service Requests ────────────────────────
...
    AORegistrationStep:
      type: object
      required:
        - step
        - data
      properties:
        step:
          type: integer
        data:
          type: object
          additionalProperties: true